+++ package.json
@@ -75,6 +75,8 @@
     "multer": "^1.4.5-lts.1",
     "multer-s3": "^3.0.1",
     "nodemailer": "^6.9.12",
+    "passport": "^0.7.0",
+    "passport-google-oauth20": "^2.0.0",
     "pino": "^8.19.0",
     "pino-pretty": "^10.3.1",
     "rate-limit-redis": "^4.0.0",
